api/requirements.txt
tensorflow==2.6.0
fastapi==0.68.0
uvicorn==0.15.0
python-multipart==0.0.5
Pillow==8.3.2
matplotlib==3.7.1
numpy==1.21.2
pydantic==1.10.8
python-dotenv==1.0.0
starlette==0.27.0
typing-extensions==4.6.3
mangum==0.17.0